Move Charge into its own file I need to split the Charge class up, but let's get this out of the way first.
