+++ uber/site_sections/api.py
@@ -13,7 +13,8 @@ from uber.config import c
 from uber.decorators import ajax, all_renderable, not_site_mappable, public, site_mappable
 from uber.errors import HTTPRedirect
 from uber.models import AdminAccount, ApiJob, ApiToken
-from uber.utils import Charge, check
+from uber.utils import check
+from uber.payments import Charge
 
 
 @all_renderable()
