+++ uber/site_sections/marketplace.py
@@ -6,7 +6,8 @@ from uber.decorators import ajax, all_renderable, render, credit_card, requires_
 from uber.errors import HTTPRedirect
 from uber.models import MarketplaceApplication
 from uber.tasks.email import send_email
-from uber.utils import Charge, check
+from uber.utils import check
+from uber.payments import Charge
 
 
 @all_renderable(public=True)
