+++ uber/site_sections/marketplace_admin.py
@@ -13,7 +13,8 @@ from uber.config import c
 from uber.decorators import ajax, all_renderable, credit_card
 from uber.errors import HTTPRedirect
 from uber.models import Attendee, Tracking, ArbitraryCharge, MarketplaceApplication
-from uber.utils import Charge, check, localized_now, Order, remove_opt
+from uber.utils import check, localized_now, Order, remove_opt
+from uber.payments import Charge
 
 
 @all_renderable()
